+++ modules/webgl/src/constants/webgl-types.ts
@@ -3,7 +3,7 @@
 // Copyright (c) vis.gl contributors
 
 /* eslint-disable camelcase */
-import {GL} from './webgl-constants';
+import {GL, type GLConstant, type GLValue} from './webgl-constants';
 
 /** Type covering all typed arrays and classic arrays consisting of numbers */
 export type NumericArray = TypedArray | number[];
@@ -30,94 +30,94 @@ type Framebuffer = unknown;
 
 /** All possible texture targets */
 export type GLTextureTarget =
-  | GL.TEXTURE_2D
-  | GL.TEXTURE_CUBE_MAP
-  | GL.TEXTURE_2D_ARRAY
-  | GL.TEXTURE_3D;
+  | GLConstant<'TEXTURE_2D'>
+  | GLConstant<'TEXTURE_CUBE_MAP'>
+  | GLConstant<'TEXTURE_2D_ARRAY'>
+  | GLConstant<'TEXTURE_3D'>;
 
 /** All possible cube face targets for textImage2D */
 export type GLTextureCubeMapTarget =
-  | GL.TEXTURE_CUBE_MAP_POSITIVE_X
-  | GL.TEXTURE_CUBE_MAP_NEGATIVE_X
-  | GL.TEXTURE_CUBE_MAP_POSITIVE_Y
-  | GL.TEXTURE_CUBE_MAP_NEGATIVE_Y
-  | GL.TEXTURE_CUBE_MAP_POSITIVE_Z
-  | GL.TEXTURE_CUBE_MAP_NEGATIVE_Z;
+  | GLConstant<'TEXTURE_CUBE_MAP_POSITIVE_X'>
+  | GLConstant<'TEXTURE_CUBE_MAP_NEGATIVE_X'>
+  | GLConstant<'TEXTURE_CUBE_MAP_POSITIVE_Y'>
+  | GLConstant<'TEXTURE_CUBE_MAP_NEGATIVE_Y'>
+  | GLConstant<'TEXTURE_CUBE_MAP_POSITIVE_Z'>
+  | GLConstant<'TEXTURE_CUBE_MAP_NEGATIVE_Z'>;
 
 /** Texel data formats for gl.texSubImage() */
 export type GLTexelDataFormat =
-  | GL.ALPHA // Discards the red, green and blue components and reads the alpha component.
-  | GL.RGB // Discards the alpha components and reads the red, green and blue components.
-  | GL.RGBA // Red, green, blue and alpha components are read from the color buffer.
-  | GL.LUMINANCE // Each color component is a luminance component, alpha is 1.0.
-  | GL.LUMINANCE_ALPHA // Each component is a luminance/alpha component.
-  | GL.SRGB
+  | GLConstant<'ALPHA'> // Discards the red, green and blue components and reads the alpha component.
+  | GLConstant<'RGB'> // Discards the alpha components and reads the red, green and blue components.
+  | GLConstant<'RGBA'> // Red, green, blue and alpha components are read from the color buffer.
+  | GLConstant<'LUMINANCE'> // Each color component is a luminance component, alpha is 1.0.
+  | GLConstant<'LUMINANCE_ALPHA'> // Each component is a luminance/alpha component.
+  | GLConstant<'SRGB'>
   // | GL.SRGB_ALPHA_EXT
-  | GL.RED
-  | GL.RG
-  | GL.RED_INTEGER
-  | GL.RG_INTEGER
-  | GL.RGB_INTEGER
-  | GL.RGBA_INTEGER
-  | GL.DEPTH_COMPONENT
-  | GL.DEPTH_STENCIL;
+  | GLConstant<'RED'>
+  | GLConstant<'RG'>
+  | GLConstant<'RED_INTEGER'>
+  | GLConstant<'RG_INTEGER'>
+  | GLConstant<'RGB_INTEGER'>
+  | GLConstant<'RGBA_INTEGER'>
+  | GLConstant<'DEPTH_COMPONENT'>
+  | GLConstant<'DEPTH_STENCIL'>;
 
 /** Rendering primitives. Constants passed to drawElements() or drawArrays() to specify what kind of primitive to render. */
 export type GLPrimitiveTopology =
-  | GL.POINTS
-  | GL.LINES
-  | GL.LINE_STRIP
-  | GL.LINE_LOOP
-  | GL.TRIANGLES
-  | GL.TRIANGLE_STRIP
-  | GL.TRIANGLE_FAN;
+  | GLConstant<'POINTS'>
+  | GLConstant<'LINES'>
+  | GLConstant<'LINE_STRIP'>
+  | GLConstant<'LINE_LOOP'>
+  | GLConstant<'TRIANGLES'>
+  | GLConstant<'TRIANGLE_STRIP'>
+  | GLConstant<'TRIANGLE_FAN'>;
 
 /** Rendering primitives. Constants passed to transform feedback  . */
-export type GLPrimitive = GL.POINTS | GL.LINES | GL.TRIANGLES;
+export type GLPrimitive = GLConstant<'POINTS'> | GLConstant<'LINES'> | GLConstant<'TRIANGLES'>;
 
 /** Data Type */
 export type GLDataType =
-  | GL.FLOAT
-  | GL.UNSIGNED_SHORT
-  | GL.UNSIGNED_INT
-  | GL.UNSIGNED_BYTE
-  | GL.BYTE
-  | GL.SHORT
-  | GL.INT
-  | GL.HALF_FLOAT;
+  | GLConstant<'FLOAT'>
+  | GLConstant<'UNSIGNED_SHORT'>
+  | GLConstant<'UNSIGNED_INT'>
+  | GLConstant<'UNSIGNED_BYTE'>
+  | GLConstant<'BYTE'>
+  | GLConstant<'SHORT'>
+  | GLConstant<'INT'>
+  | GLConstant<'HALF_FLOAT'>;
 
 /** Pixel Data Type */
 export type GLPixelType =
   | GLDataType
-  | GL.UNSIGNED_SHORT_5_6_5
-  | GL.UNSIGNED_SHORT_4_4_4_4
-  | GL.UNSIGNED_SHORT_5_5_5_1
-  | GL.UNSIGNED_INT_2_10_10_10_REV
-  | GL.UNSIGNED_INT_10F_11F_11F_REV
-  | GL.UNSIGNED_INT_5_9_9_9_REV
-  | GL.UNSIGNED_INT_24_8
-  | GL.FLOAT_32_UNSIGNED_INT_24_8_REV;
+  | GLConstant<'UNSIGNED_SHORT_5_6_5'>
+  | GLConstant<'UNSIGNED_SHORT_4_4_4_4'>
+  | GLConstant<'UNSIGNED_SHORT_5_5_5_1'>
+  | GLConstant<'UNSIGNED_INT_2_10_10_10_REV'>
+  | GLConstant<'UNSIGNED_INT_10F_11F_11F_REV'>
+  | GLConstant<'UNSIGNED_INT_5_9_9_9_REV'>
+  | GLConstant<'UNSIGNED_INT_24_8'>
+  | GLConstant<'FLOAT_32_UNSIGNED_INT_24_8_REV'>;
 
 /**
  * Sampler uniform type
  * @note These are all the valid sampler types used with `gl.uniform1i((location, value)`
  */
 export type GLSamplerType =
-  | GL.SAMPLER_2D
-  | GL.SAMPLER_CUBE
-  | GL.SAMPLER_3D
-  | GL.SAMPLER_2D_SHADOW
-  | GL.SAMPLER_2D_ARRAY
-  | GL.SAMPLER_2D_ARRAY_SHADOW
-  | GL.SAMPLER_CUBE_SHADOW
-  | GL.INT_SAMPLER_2D
-  | GL.INT_SAMPLER_3D
-  | GL.INT_SAMPLER_CUBE
-  | GL.INT_SAMPLER_2D_ARRAY
-  | GL.UNSIGNED_INT_SAMPLER_2D
-  | GL.UNSIGNED_INT_SAMPLER_3D
-  | GL.UNSIGNED_INT_SAMPLER_CUBE
-  | GL.UNSIGNED_INT_SAMPLER_2D_ARRAY;
+  | GLConstant<'SAMPLER_2D'>
+  | GLConstant<'SAMPLER_CUBE'>
+  | GLConstant<'SAMPLER_3D'>
+  | GLConstant<'SAMPLER_2D_SHADOW'>
+  | GLConstant<'SAMPLER_2D_ARRAY'>
+  | GLConstant<'SAMPLER_2D_ARRAY_SHADOW'>
+  | GLConstant<'SAMPLER_CUBE_SHADOW'>
+  | GLConstant<'INT_SAMPLER_2D'>
+  | GLConstant<'INT_SAMPLER_3D'>
+  | GLConstant<'INT_SAMPLER_CUBE'>
+  | GLConstant<'INT_SAMPLER_2D_ARRAY'>
+  | GLConstant<'UNSIGNED_INT_SAMPLER_2D'>
+  | GLConstant<'UNSIGNED_INT_SAMPLER_3D'>
+  | GLConstant<'UNSIGNED_INT_SAMPLER_CUBE'>
+  | GLConstant<'UNSIGNED_INT_SAMPLER_2D_ARRAY'>;
 
 /**
  * Composite types table
@@ -125,107 +125,121 @@ export type GLSamplerType =
  * Different `gl.uniformXXX(location, value)` functions must be used depending on which composite type is being set.
  */
 export type GLUniformType =
-  | GL.FLOAT
-  | GL.FLOAT_VEC2
-  | GL.FLOAT_VEC3
-  | GL.FLOAT_VEC4
-  | GL.INT
-  | GL.INT_VEC2
-  | GL.INT_VEC3
-  | GL.INT_VEC4
-  | GL.UNSIGNED_INT
-  | GL.UNSIGNED_INT_VEC2
-  | GL.UNSIGNED_INT_VEC3
-  | GL.UNSIGNED_INT_VEC4
-  | GL.BOOL
-  | GL.BOOL_VEC2
-  | GL.BOOL_VEC3
-  | GL.BOOL_VEC4
-  | GL.FLOAT_MAT2
-  | GL.FLOAT_MAT2x3
-  | GL.FLOAT_MAT2x4
-  | GL.FLOAT_MAT3x2
-  | GL.FLOAT_MAT3
-  | GL.FLOAT_MAT3x4
-  | GL.FLOAT_MAT4x2
-  | GL.FLOAT_MAT4x3
-  | GL.FLOAT_MAT4;
+  | GLConstant<'FLOAT'>
+  | GLConstant<'FLOAT_VEC2'>
+  | GLConstant<'FLOAT_VEC3'>
+  | GLConstant<'FLOAT_VEC4'>
+  | GLConstant<'INT'>
+  | GLConstant<'INT_VEC2'>
+  | GLConstant<'INT_VEC3'>
+  | GLConstant<'INT_VEC4'>
+  | GLConstant<'UNSIGNED_INT'>
+  | GLConstant<'UNSIGNED_INT_VEC2'>
+  | GLConstant<'UNSIGNED_INT_VEC3'>
+  | GLConstant<'UNSIGNED_INT_VEC4'>
+  | GLConstant<'BOOL'>
+  | GLConstant<'BOOL_VEC2'>
+  | GLConstant<'BOOL_VEC3'>
+  | GLConstant<'BOOL_VEC4'>
+  | GLConstant<'FLOAT_MAT2'>
+  | GLConstant<'FLOAT_MAT2x3'>
+  | GLConstant<'FLOAT_MAT2x4'>
+  | GLConstant<'FLOAT_MAT3x2'>
+  | GLConstant<'FLOAT_MAT3'>
+  | GLConstant<'FLOAT_MAT3x4'>
+  | GLConstant<'FLOAT_MAT4x2'>
+  | GLConstant<'FLOAT_MAT4x3'>
+  | GLConstant<'FLOAT_MAT4'>;
 
 /**
  * Depth or stencil tests
  * Constants passed to WebGLRenderingContext.depthFunc() or WebGLRenderingContext.stencilFunc().
  */
 export type GLFunction =
-  | GL.NEVER
-  | GL.LESS
-  | GL.EQUAL
-  | GL.LEQUAL
-  | GL.GREATER
-  | GL.NOTEQUAL
-  | GL.GEQUAL
-  | GL.ALWAYS;
+  | GLConstant<'NEVER'>
+  | GLConstant<'LESS'>
+  | GLConstant<'EQUAL'>
+  | GLConstant<'LEQUAL'>
+  | GLConstant<'GREATER'>
+  | GLConstant<'NOTEQUAL'>
+  | GLConstant<'GEQUAL'>
+  | GLConstant<'ALWAYS'>;
 
 export type GLBlendEquation =
-  | GL.FUNC_ADD
-  | GL.FUNC_SUBTRACT
-  | GL.FUNC_REVERSE_SUBTRACT
-  | GL.MIN
-  | GL.MAX;
+  | GLConstant<'FUNC_ADD'>
+  | GLConstant<'FUNC_SUBTRACT'>
+  | GLConstant<'FUNC_REVERSE_SUBTRACT'>
+  | GLConstant<'MIN'>
+  | GLConstant<'MAX'>;
 
 export type GLBlendFunction =
-  | GL.ZERO
-  | GL.ONE
-  | GL.SRC_COLOR
-  | GL.ONE_MINUS_SRC_COLOR
-  | GL.DST_COLOR
-  | GL.ONE_MINUS_DST_COLOR
-  | GL.SRC_ALPHA
-  | GL.ONE_MINUS_SRC_ALPHA
-  | GL.DST_ALPHA
-  | GL.ONE_MINUS_DST_ALPHA
-  | GL.CONSTANT_COLOR
-  | GL.ONE_MINUS_CONSTANT_COLOR
-  | GL.CONSTANT_ALPHA
-  | GL.ONE_MINUS_CONSTANT_ALPHA
-  | GL.SRC_ALPHA_SATURATE;
+  | GLConstant<'ZERO'>
+  | GLConstant<'ONE'>
+  | GLConstant<'SRC_COLOR'>
+  | GLConstant<'ONE_MINUS_SRC_COLOR'>
+  | GLConstant<'DST_COLOR'>
+  | GLConstant<'ONE_MINUS_DST_COLOR'>
+  | GLConstant<'SRC_ALPHA'>
+  | GLConstant<'ONE_MINUS_SRC_ALPHA'>
+  | GLConstant<'DST_ALPHA'>
+  | GLConstant<'ONE_MINUS_DST_ALPHA'>
+  | GLConstant<'CONSTANT_COLOR'>
+  | GLConstant<'ONE_MINUS_CONSTANT_COLOR'>
+  | GLConstant<'CONSTANT_ALPHA'>
+  | GLConstant<'ONE_MINUS_CONSTANT_ALPHA'>
+  | GLConstant<'SRC_ALPHA_SATURATE'>;
 
 /**
  * Stencil actions
  * Constants passed to WebGLRenderingContext.stencilOp().
  */
 export type GLStencilOp =
-  | GL.KEEP
-  | GL.ZERO
-  | GL.REPLACE
-  | GL.INCR
-  | GL.INCR_WRAP
-  | GL.DECR
-  | GL.DECR_WRAP
-  | GL.INVERT;
+  | GLConstant<'KEEP'>
+  | GLConstant<'ZERO'>
+  | GLConstant<'REPLACE'>
+  | GLConstant<'INCR'>
+  | GLConstant<'INCR_WRAP'>
+  | GLConstant<'DECR'>
+  | GLConstant<'DECR_WRAP'>
+  | GLConstant<'INVERT'>;
 
-export type GLPolygonMode = GL.FILL_WEBGL | GL.LINE_WEBGL;
-export type GLCullFaceMode = GL.FRONT | GL.BACK | GL.FRONT_AND_BACK;
-export type GLProvokingVertex = GL.FIRST_VERTEX_CONVENTION_WEBGL | GL.LAST_VERTEX_CONVENTION_WEBGL;
+export type GLPolygonMode = GLConstant<'FILL_WEBGL'> | GLConstant<'LINE_WEBGL'>;
+export type GLCullFaceMode =
+  | GLConstant<'FRONT'>
+  | GLConstant<'BACK'>
+  | GLConstant<'FRONT_AND_BACK'>;
+export type GLProvokingVertex =
+  | GLConstant<'FIRST_VERTEX_CONVENTION_WEBGL'>
+  | GLConstant<'LAST_VERTEX_CONVENTION_WEBGL'>;
 
 /** Parameters for textures and samplers */
 export type GLSamplerParameters = {
   /** Sets the wrap parameter for texture coordinate  to either GL_CLAMP_TO_EDGE, GL_MIRRORED_REPEAT, or GL_REPEAT. */
-  [GL.TEXTURE_WRAP_S]?: GL.CLAMP_TO_EDGE | GL.REPEAT | GL.MIRRORED_REPEAT;
+  [GL.TEXTURE_WRAP_S]?:
+    | GLConstant<'CLAMP_TO_EDGE'>
+    | GLConstant<'REPEAT'>
+    | GLConstant<'MIRRORED_REPEAT'>;
   /** Sets the wrap parameter for texture coordinate  to either GL_CLAMP_TO_EDGE, GL_MIRRORED_REPEAT, or GL_REPEAT. */
-  [GL.TEXTURE_WRAP_T]?: GL.CLAMP_TO_EDGE | GL.REPEAT | GL.MIRRORED_REPEAT;
+  [GL.TEXTURE_WRAP_T]?:
+    | GLConstant<'CLAMP_TO_EDGE'>
+    | GLConstant<'REPEAT'>
+    | GLConstant<'MIRRORED_REPEAT'>;
   /** Sets the wrap parameter for texture coordinate  to either GL_CLAMP_TO_EDGE, GL_MIRRORED_REPEAT, or GL_REPEAT. */
-  [GL.TEXTURE_WRAP_R]?: GL.CLAMP_TO_EDGE | GL.REPEAT | GL.MIRRORED_REPEAT;
+  [GL.TEXTURE_WRAP_R]?:
+    | GLConstant<'CLAMP_TO_EDGE'>
+    | GLConstant<'REPEAT'>
+    | GLConstant<'MIRRORED_REPEAT'>;
 
   /** The texture magnification function is used when the pixel being textured maps to an area less than or equal to one texture element. It sets the texture magnification function to either GL_NEAREST or GL_LINEAR (see below). GL_NEAREST is generally faster than GL_LINEAR, but it can produce textured images with sharper edges because the transition between texture elements is not as smooth. Default: GL_LINEAR.  */
-  [GL.TEXTURE_MAG_FILTER]?: GL.NEAREST | GL.LINEAR;
+  [GL.TEXTURE_MAG_FILTER]?: GLConstant<'NEAREST'> | GLConstant<'LINEAR'>;
   /** The texture minifying function is used whenever the pixel being textured maps to an area greater than one texture element. There are six defined minifying functions. Two of them use the nearest one or nearest four texture elements to compute the texture value. The other four use mipmaps. Default: GL_NEAREST_MIPMAP_LINEAR */
   [GL.TEXTURE_MIN_FILTER]?:
-    | GL.NEAREST
-    | GL.LINEAR
-    | GL.NEAREST_MIPMAP_NEAREST
-    | GL.NEAREST_MIPMAP_LINEAR
-    | GL.LINEAR_MIPMAP_NEAREST
-    | GL.LINEAR_MIPMAP_LINEAR;
+    | GLConstant<'NEAREST'>
+    | GLConstant<'LINEAR'>
+    | GLConstant<'NEAREST_MIPMAP_NEAREST'>
+    | GLConstant<'NEAREST_MIPMAP_LINEAR'>
+    | GLConstant<'LINEAR_MIPMAP_NEAREST'>
+    | GLConstant<'LINEAR_MIPMAP_LINEAR'>;
   /* A GLfloat indicating the minimum level-of-detail mipmap. */
   [GL.TEXTURE_MIN_LOD]?: number;
   /* A GLfloat indicating the minimum level-of-detail mipmap. */
@@ -233,7 +247,7 @@ export type GLSamplerParameters = {
   /** Texture parameter TEXTURE_COMPARE_FUNC specifies the depth texture comparison function */
   [GL.TEXTURE_COMPARE_FUNC]?: number; // COMPARE_FUNC);
   /** Texture parameter TEXTURE_COMPARE_MODE specifies the depth texture comparison operands. */
-  [GL.TEXTURE_COMPARE_MODE]?: GL.COMPARE_REF_TO_TEXTURE;
+  [GL.TEXTURE_COMPARE_MODE]?: GLConstant<'COMPARE_REF_TO_TEXTURE'>;
   /** Max anisotropy level */
   [GL.TEXTURE_MAX_ANISOTROPY_EXT]?: number;
 };
@@ -253,14 +267,17 @@ export type GLValueParameters = {
   [GL.COLOR_CLEAR_VALUE]?: [number, number, number, number] | TypedArray;
   [GL.COLOR_WRITEMASK]?: [boolean, boolean, boolean, boolean] | boolean[];
   [GL.CULL_FACE]?: boolean;
-  [GL.CULL_FACE_MODE]?: GL.FRONT | GL.BACK | GL.FRONT_AND_BACK;
+  [GL.CULL_FACE_MODE]?: GLConstant<'FRONT'> | GLConstant<'BACK'> | GLConstant<'FRONT_AND_BACK'>;
   [GL.DEPTH_TEST]?: boolean;
   [GL.DEPTH_CLEAR_VALUE]?: number;
   [GL.DEPTH_FUNC]?: GLFunction;
   [GL.DEPTH_RANGE]?: [number, number] | TypedArray;
   [GL.DEPTH_WRITEMASK]?: boolean;
   [GL.DITHER]?: boolean;
-  [GL.FRAGMENT_SHADER_DERIVATIVE_HINT]?: GL.FASTEST | GL.NICEST | GL.DONT_CARE;
+  [GL.FRAGMENT_SHADER_DERIVATIVE_HINT]?:
+    | GLConstant<'FASTEST'>
+    | GLConstant<'NICEST'>
+    | GLConstant<'DONT_CARE'>;
   [GL.CURRENT_PROGRAM]?: WebGLProgram | null;
   [GL.FRAMEBUFFER_BINDING]?: WebGLFramebuffer | null;
   [GL.RENDERBUFFER_BINDING]?: WebGLRenderbuffer | null;
@@ -275,8 +292,11 @@ export type GLValueParameters = {
   [GL.TEXTURE_BINDING_2D_ARRAY]?: WebGLTexture | null;
   [GL.TEXTURE_BINDING_3D]?: WebGLTexture | null;
   [GL.TEXTURE_BINDING_CUBE_MAP]?: WebGLTexture | null;
-  [GL.FRONT_FACE]?: GL.CW | GL.CCW;
-  [GL.GENERATE_MIPMAP_HINT]?: GL.FASTEST | GL.NICEST | GL.DONT_CARE;
+  [GL.FRONT_FACE]?: GLConstant<'CW'> | GLConstant<'CCW'>;
+  [GL.GENERATE_MIPMAP_HINT]?:
+    | GLConstant<'FASTEST'>
+    | GLConstant<'NICEST'>
+    | GLConstant<'DONT_CARE'>;
   [GL.LINE_WIDTH]?: number;
   [GL.POLYGON_OFFSET_FILL]?: boolean;
   [GL.POLYGON_OFFSET_FACTOR]?: number;
@@ -316,7 +336,9 @@ export type GLValueParameters = {
   [GL.UNPACK_ALIGNMENT]?: 1 | 2 | 4 | 8;
   [GL.UNPACK_FLIP_Y_WEBGL]?: boolean;
   [GL.UNPACK_PREMULTIPLY_ALPHA_WEBGL]?: boolean;
-  [GL.UNPACK_COLORSPACE_CONVERSION_WEBGL]?: GL.NONE | GL.BROWSER_DEFAULT_WEBGL;
+  [GL.UNPACK_COLORSPACE_CONVERSION_WEBGL]?:
+    | GLConstant<'NONE'>
+    | GLConstant<'BROWSER_DEFAULT_WEBGL'>;
   [GL.UNPACK_ROW_LENGTH]?: number;
   [GL.UNPACK_IMAGE_HEIGHT]?: number;
   [GL.UNPACK_SKIP_PIXELS]?: number;
@@ -335,7 +357,9 @@ export type GLUnpackParameters = {
   [GL.UNPACK_ALIGNMENT]?: number;
   [GL.UNPACK_FLIP_Y_WEBGL]?: boolean;
   [GL.UNPACK_PREMULTIPLY_ALPHA_WEBGL]?: boolean;
-  [GL.UNPACK_COLORSPACE_CONVERSION_WEBGL]?: GL.NONE | GL.BROWSER_DEFAULT_WEBGL;
+  [GL.UNPACK_COLORSPACE_CONVERSION_WEBGL]?:
+    | GLConstant<'NONE'>
+    | GLConstant<'BROWSER_DEFAULT_WEBGL'>;
   [GL.UNPACK_ROW_LENGTH]?: number;
   [GL.UNPACK_IMAGE_HEIGHT]?: number;
   [GL.UNPACK_SKIP_PIXELS]?: number;
@@ -365,7 +389,7 @@ export type GLFunctionParameters = {
   colorMask?: [boolean, boolean, boolean, boolean] | boolean[];
 
   cull?: boolean;
-  cullFace?: GL.FRONT | GL.BACK | GL.FRONT_AND_BACK;
+  cullFace?: GLConstant<'FRONT'> | GLConstant<'BACK'> | GLConstant<'FRONT_AND_BACK'>;
 
   depthTest?: boolean;
   depthFunc?: GLFunction;
@@ -375,11 +399,11 @@ export type GLFunctionParameters = {
 
   dither?: boolean;
 
-  derivativeHint?: GL.FASTEST | GL.NICEST | GL.DONT_CARE;
+  derivativeHint?: GLConstant<'FASTEST'> | GLConstant<'NICEST'> | GLConstant<'DONT_CARE'>;
 
-  frontFace?: GL.CW | GL.CCW;
+  frontFace?: GLConstant<'CW'> | GLConstant<'CCW'>;
 
-  mipmapHint?: GL.FASTEST | GL.NICEST | GL.DONT_CARE;
+  mipmapHint?: GLConstant<'FASTEST'> | GLConstant<'NICEST'> | GLConstant<'DONT_CARE'>;
 
   lineWidth?: number;
 
@@ -608,14 +632,19 @@ type WEBGL_provoking_vertex = {
   // Constants in GL enum
   /** Set the provoking vertex */
   provokingVertexWEBGL(
-    provokeMode: GL.FIRST_VERTEX_CONVENTION_WEBGL | GL.LAST_VERTEX_CONVENTION_WEBGL
+    provokeMode:
+      | GLConstant<'FIRST_VERTEX_CONVENTION_WEBGL'>
+      | GLConstant<'LAST_VERTEX_CONVENTION_WEBGL'>
   ): void;
 };
 
 /** WEBGL_polygon_mode https://registry.khronos.org/webgl/extensions/WEBGL_polygon_mode/ */
 type WEBGL_polygon_mode = {
   /** Set polygon mode of face to fill or line */
-  polygonModeWEBGL(face: GL.FRONT | GL.BACK, mode: GL.LINE_WEBGL | GL.FILL_WEBGL): void;
+  polygonModeWEBGL(
+    face: GLConstant<'FRONT'> | GLConstant<'BACK'>,
+    mode: GLConstant<'LINE_WEBGL'> | GLConstant<'FILL_WEBGL'>
+  ): void;
 };
 
 /** WEBGL_clip_cull_distance https://registry.khronos.org/webgl/extensions/WEBGL_clip_cull_distance/ */
@@ -672,7 +701,7 @@ type EXT_clip_control = {
   CLIP_ORIGIN_EXT: 0x935c;
   CLIP_DEPTH_MODE_EXT: 0x935d;
 
-  clipControlEXT(origin: GL, depth: GL): void;
+  clipControlEXT(origin: GLValue, depth: GLValue): void;
 };
 
 /** WEBGL_blend_func_extended https://registry.khronos.org/webgl/extensions/WEBGL_blend_func_extended/ */
@@ -687,17 +716,23 @@ type WEBGL_blend_func_extended = {
 /** OES_draw_buffers_indexed https://registry.khronos.org/webgl/extensions/OES_draw_buffers_indexed/ */
 type OES_draw_buffers_indexed = {
   /** Enables blending for an individual draw buffer */
-  enableiOES(target: GL, index: number): void;
+  enableiOES(target: GLValue, index: number): void;
   /** Disables blending for an individual draw buffer */
-  disableiOES(target: GL, index: number): void;
+  disableiOES(target: GLValue, index: number): void;
   /** Modifies blend equation for an individual draw buffer */
-  blendEquationiOES(buf: number, mode: GL): void;
+  blendEquationiOES(buf: number, mode: GLValue): void;
   /** Modifies blend equation for an individual draw buffer */
-  blendEquationSeparateiOES(buf: number, modeRGB: GL, modeAlpha: GL): void;
+  blendEquationSeparateiOES(buf: number, modeRGB: GLValue, modeAlpha: GLValue): void;
   /** Modifies blend function for an individual draw buffer */
-  blendFunciOES(buf: number, src: GL, dst: GL): void;
+  blendFunciOES(buf: number, src: GLValue, dst: GLValue): void;
   /** Modifies blend function for an individual draw buffer */
-  blendFuncSeparateiOES(buf: number, srcRGB: GL, dstRGB: GL, srcAlpha: GL, dstAlpha: GL): void;
+  blendFuncSeparateiOES(
+    buf: number,
+    srcRGB: GLValue,
+    dstRGB: GLValue,
+    srcAlpha: GLValue,
+    dstAlpha: GLValue
+  ): void;
   /** Modifies color mask for an individual draw buffer */
   colorMaskiOES(buf: number, r: boolean, g: boolean, b: boolean, a: boolean): void;
 };
@@ -806,8 +841,8 @@ type WEBGL_shader_pixel_local_storage = {
   // framebufferPixelLocalClearValueuivWEBGL(plane: number,
   //                                                   Uint32List value,
   //                                                   optional unsigned long long srcOffset = 0): void;
-  beginPixelLocalStorageWEBGL(loadops: GL[]): void;
-  endPixelLocalStorageWEBGL(storeops: GL[]): void;
+  beginPixelLocalStorageWEBGL(loadops: GLValue[]): void;
+  endPixelLocalStorageWEBGL(storeops: GLValue[]): void;
   pixelLocalStorageBarrierWEBGL(): void;
-  getFramebufferPixelLocalStorageParameterWEBGL(plane: number, pname: GL): any;
+  getFramebufferPixelLocalStorageParameterWEBGL(plane: number, pname: GLValue): any;
 };