+++ modules/webgl/src/webgl-constants.ts
@@ -3,3 +3,4 @@
 // Copyright (c) vis.gl contributors
 
 export {GL} from './constants/webgl-constants';
+export type {GLConstant, GLValue} from './constants/webgl-constants';